% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pool.R
\name{create_pool}
\alias{create_pool}
\title{Create a pool of connections}
\usage{
create_pool(options_list)
}
\arguments{
\item{options_list}{A list of options for connecting to a database, usually generated by calling \code{credentials_extract()}}
}
\description{
Use a list of options (usually created through running \code{credentials_extract})
to create a pool of connections to the psql database.
}
