% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio_indicators_db.R
\name{create_portfolio_indicators_db}
\alias{create_portfolio_indicators_db}
\title{Create an initial portfolio_indicators table in the psql database}
\usage{
create_portfolio_indicators_db(portfolio_indicators = NULL,
  connection_object = NULL)
}
\arguments{
\item{portfolio_indicators}{A dataframe with the columns portfolio_id, indicator_id}

\item{connection_object}{An open connection to adatabase (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}
}
\value{
A "portfolio_indicators" table added to PSQL database
}
\description{
Create an initial portfolio_users data table in the psql database
}
