% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio_users_db.R
\name{create_portfolio_users_db}
\alias{create_portfolio_users_db}
\title{Create an initial portfolio_users data table in the psql database}
\usage{
create_portfolio_users_db(portfolio_users = NULL, connection_object = NULL)
}
\arguments{
\item{portfolio_users}{A dataframe with the columns portfolio_id, user_id, is_creator, and is_admin}

\item{connection_object}{An open connection to adatabase (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}
}
\value{
A "portfolio_users" table added to PSQL database
}
\description{
Create an initial portfolio_users data table in the psql database
}
