% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolios_db.R
\name{create_portfolios_db}
\alias{create_portfolios_db}
\title{Create an initial portfolios table in the psql database}
\usage{
create_portfolios_db(portfolios = NULL, connection_object = NULL)
}
\arguments{
\item{portfolios}{A dataframe with the columns portfolio_id, portfolio_name}

\item{connection_object}{An open connection to adatabase (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}
}
\value{
A "portfolios" table added to PSQL database
}
\description{
Create an initial portfolio_users data table in the psql database
}
