% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user_portfolio_data.R
\name{create_user_portfolio_data}
\alias{create_user_portfolio_data}
\title{Create an initial user_portfolio data table in the psql database}
\usage{
create_user_portfolio_data(users = NULL, connection_object = NULL)
}
\arguments{
\item{users}{A SQL query as a chracter string of length 1}

\item{connection_object}{An open connection to adatabase (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}
}
\value{
A "user_portfolio" table added to PSQL database
}
\description{
Fetch a table or a specific query of a table from the database.
}
