% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_connect.R
\name{credentials_connect}
\alias{credentials_connect}
\title{Create a connection to database}
\usage{
credentials_connect(options_list)
}
\arguments{
\item{options_list}{A list of options for connecting to a database, usually generated by calling \code{credentials_extract()}}
}
\description{
Use a list of options (usually created through running \code{credentials_extract})
to create a connection to the psql database.
}
