% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_extract.R
\name{credentials_extract}
\alias{credentials_extract}
\title{Extract information from a credentials file}
\usage{
credentials_extract(credentials_path = NULL, credentials_folder = TRUE,
  credentials_file = "credentials.yaml", credentials_search_limit = 10,
  all_in_file = FALSE)
}
\arguments{
\item{credentials_path}{The path to a credentials file through which
the database connection will be established.
If \code{NULL} the function will attempt to find an \code{credentials.yaml}
upwards in the directory tree}

\item{credentials_folder}{Whether to search for a \code{credentials} folder (the
default behvior); only applicable if \code{credentials_path} is \code{NULL}.}

\item{credentials_file}{The name of the file (not including the path, unless
\code{all_in_file} is set to \code{TRUE}) with
your environment setting. Normally, this should be \code{credentials.yaml}. If an explicit \code{credentials_path} is provided, this
argument is ignored. Otherwise, the \code{credentials_file} will be
searched for, with the expectation that it is in a directory
named \code{credentials}.}

\item{credentials_search_limit}{The number of directories upwards to search for a \code{credentials/credentials.yaml} file; only relevant if \code{credentials_path = NULL}
default is \code{FALSE}}

\item{all_in_file}{Whether to disregard the \code{credentials_path} argument entirely.
If set to \code{TRUE}, the \code{credentials_file} argument should be the
entire pathway to the credentials file in question (ie, path and file name)
combined.
If set to \code{TRUE}, the function will use the}
}
\value{
A list of PostgreSQL-related options obtained from the
environment file, meant to be passed  to \code{credentials_connect}.
}
\description{
Extract information from a \code{credetials.yaml} so as to access a database.
This will generally used in conjunction with \code{credentials_connect}
The list generated by \code{credentials_extract} is passed as the \code{options_list} argument of \code{credentials_connect}).
}
