% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_portfolio.R
\name{filter_portfolio}
\alias{filter_portfolio}
\title{Filter portfolio}
\usage{
filter_portfolio(portfolio, ...)
}
\value{
A filtered as_portfolio table
}
\description{
Filter an as_portfolio table with any number of filter conditions. Each condition is a separate argument, wrapped in quotations.
}
\examples{
fp <- filter_portfolio(portfolio = as_portfolio)
fp <- filter_portfolio(portfolio = as_portfolio, 'ifc_investee_pct == 100', 'project_name != "SI Aureos GTAA"')
fp <- filter_portfolio(portfolio = as_portfolio, 'ifc_investee_pct == 100', 'project_name != "SI Aureos GTAA"')
}
