% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from database}
\usage{
get_data(query = NULL, tab = NULL, dbname = NULL, host = NULL,
  port = NULL, user = NULL, password = NULL, connection_object = NULL,
  collect = TRUE)
}
\arguments{
\item{query}{A SQL query as a chracter string of length 1}

\item{tab}{As an alternative to query, the name of an entire table to return}

\item{dbname}{The name of the database from which you are getting data.
Example: portfolio. If \code{NULL}
the function will try to use the \code{dbname} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{host}{The name of the host from which you are getting data.
If \code{NULL} the function will try to use the \code{host} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{port}{The name of the port from which you are getting data.
If \code{NULL}
the function will try to use the \code{port} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{user}{The user. If \code{NULL}
the function will try to use the \code{user} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{password}{The password If \code{NULL}
the function will try to use the \code{pqssword} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{connection_object}{An open connection to a CISM database (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}

\item{collect}{Whether to collect into memory (using dplyr's \code{connect}
function) the result or not}
}
\value{
A dataframe matching the results of either the \code{query} or \code{tab} arguments
}
\description{
Fetch a table or a specific query of a table from the database.
}
