% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_as_results.R
\name{populate_as_results}
\alias{populate_as_results}
\title{Populate the as_results relation}
\usage{
populate_as_results(csv = "data/portfolio.as_results.csv", the_co = NULL)
}
\arguments{
\item{csv}{The location of the "portfolio.as_results.csv" file, emailed by Soren}

\item{the_co}{An open connection to adatabase (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}
}
\value{
The local "portfolios.as_results" relation in the postgresql database will be updated
}
\description{
Populate the as_results relation from "portfolio.as_results.csv" (emailed to developer)
}
\examples{
2+2
}
