% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Prettify a table for HTML documents}
\usage{
prettify(the_table, remove_underscores_columns = TRUE, cap_columns = TRUE,
  cap_characters = TRUE, comma_numbers = TRUE,
  date_format = "\%B \%d, \%Y", round_digits = 2, remove_row_names = TRUE,
  remove_line_breaks = TRUE, data_table = TRUE, nrows = 5,
  download_options = FALSE, no_scroll = TRUE, scroll_x = TRUE)
}
\arguments{
\item{the_table}{The dataframe to be prettified}

\item{remove_underscores_columns}{Whether to remove underscores in column names and replace them with spaces}

\item{cap_columns}{Whether to capitalize the first letter of the names of columns}

\item{cap_characters}{Whether to capitalize the first letter of elements of character vectors}

\item{comma_numbers}{Whether to include a comma between every three digits of numeric columns}

\item{date_format}{The format for printing date columns}

\item{round_digits}{How many digits to round numberic columns to}

\item{remove_row_names}{Whether to remove row names}

\item{remove_line_breaks}{Whether to remove line breaks from the elements of columns}

\item{nrows}{The number of rows to show}

\item{download_options}{Whether to show options for downloading the table}
}
\value{
A data table ready for inclusion in an .Rmd to be knitted to HTML
}
\description{
Create a data table object from a dataframe with optional aesthetic improvements. Userful for inclusion in .Rmd files being knitted to HTML
}
